Fix bash variable checking

+++ frameworks/Ruby/padrino/install.sh
@@ -5,9 +5,10 @@ fw_depends rvm nginx java
 rvm install rbx-2.2.10
 # Some gems seem to fail when installing on travis
 # clearing at the cache so it always retrieves new ones
-[ "$TRAVIS" == "true" ] || { \
+if [ "$TRAVIS" == "true" ]
+then
   rm -rf /home/travis/.rvm/gems/rbx-2.2.10/cache
-}
+fi
 rvm rbx-2.2.10 do bundle install --jobs=1 --retry=3 --gemfile=$TROOT/Gemfile
 
 rvm install ruby-2.0.0-p0
